refactor(api): deprecate POST /schooling-registration-dependent-users
in favor of POST /sco-organization-learners/dependent
