+++ api/tests/acceptance/application/organization-learner-dependent-user-controller_test.js
@@ -20,177 +20,6 @@ describe('Acceptance | Route | organization-learner-dependent-user', function ()
     await knex('authentication-methods').delete();
   });
 
-  describe('POST /api/schooling-registration-dependent-users', function () {
-    let organization;
-    let campaign;
-    let organizationLearner;
-
-    beforeEach(async function () {
-      // given
-      organization = databaseBuilder.factory.buildOrganization();
-      organizationLearner = databaseBuilder.factory.buildOrganizationLearner({
-        organizationId: organization.id,
-        userId: null,
-        nationalStudentId: 'salut',
-      });
-      campaign = databaseBuilder.factory.buildCampaign({ organizationId: organization.id });
-      await databaseBuilder.commit();
-    });
-
-    context('when creation is with email', function () {
-      it('should return an 204 status after having successfully created user and associated user to organizationLearner', async function () {
-        // when
-        const response = await server.inject({
-          method: 'POST',
-          url: '/api/schooling-registration-dependent-users',
-          payload: {
-            data: {
-              attributes: {
-                'campaign-code': campaign.code,
-                'first-name': organizationLearner.firstName,
-                'last-name': organizationLearner.lastName,
-                birthdate: organizationLearner.birthdate,
-                password: 'P@ssw0rd',
-                email: 'angie@example.net',
-                'with-username': false,
-              },
-            },
-          },
-        });
-
-        // then
-        expect(response.statusCode).to.equal(204);
-      });
-
-      context('when no organizationLearner not linked yet found', function () {
-        it('should respond with a 409 - Conflict', async function () {
-          // given
-          const userId = databaseBuilder.factory.buildUser().id;
-          const organizationLearnerAlreadyLinked = databaseBuilder.factory.buildOrganizationLearner({
-            firstName: 'josé',
-            lastName: 'bové',
-            birthdate: '2020-01-01',
-            organizationId: organization.id,
-            userId,
-            nationalStudentId: 'coucou',
-          });
-          await databaseBuilder.commit();
-
-          // when
-          const response = await server.inject({
-            method: 'POST',
-            url: '/api/schooling-registration-dependent-users',
-            payload: {
-              data: {
-                attributes: {
-                  'campaign-code': campaign.code,
-                  'first-name': organizationLearnerAlreadyLinked.firstName,
-                  'last-name': organizationLearnerAlreadyLinked.lastName,
-                  birthdate: organizationLearnerAlreadyLinked.birthdate,
-                  password: 'P@ssw0rd',
-                  email: 'angie@example.net',
-                  'with-username': false,
-                },
-              },
-            },
-          });
-
-          // then
-          expect(response.statusCode).to.equal(409);
-          expect(response.result.errors[0].detail).to.equal(
-            'Un compte existe déjà pour l‘élève dans le même établissement.'
-          );
-        });
-      });
-
-      context('when a field is not valid', function () {
-        it('should respond with a 422 - Unprocessable Entity', async function () {
-          // when
-          const response = await server.inject({
-            method: 'POST',
-            url: '/api/schooling-registration-dependent-users',
-            payload: {
-              data: {
-                attributes: {
-                  'campaign-code': campaign.code,
-                  'first-name': organizationLearner.firstName,
-                  'last-name': organizationLearner.lastName,
-                  birthdate: organizationLearner.birthdate,
-                  password: 'P@ssw0rd',
-                  email: 'not valid email',
-                  'with-username': false,
-                },
-              },
-            },
-          });
-
-          // then
-          expect(response.statusCode).to.equal(422);
-        });
-      });
-    });
-
-    context('when creation is with username', function () {
-      it('should return a 204 status after having successfully created user and associated user to organizationLearner', async function () {
-        // when
-        const response = await server.inject({
-          method: 'POST',
-          url: '/api/schooling-registration-dependent-users',
-          payload: {
-            data: {
-              attributes: {
-                'campaign-code': campaign.code,
-                'first-name': organizationLearner.firstName,
-                'last-name': organizationLearner.lastName,
-                birthdate: organizationLearner.birthdate,
-                password: 'P@ssw0rd',
-                username: 'angie.go1234',
-                'with-username': true,
-              },
-            },
-          },
-        });
-
-        // then
-        expect(response.statusCode).to.equal(204);
-      });
-
-      context('when username is already taken', function () {
-        it('should respond with a 422 - Unprocessable entity', async function () {
-          // given
-          const username = 'angie.go1234';
-          databaseBuilder.factory.buildUser({ username });
-          await databaseBuilder.commit();
-
-          // when
-          const response = await server.inject({
-            method: 'POST',
-            url: '/api/schooling-registration-dependent-users',
-            payload: {
-              data: {
-                attributes: {
-                  'campaign-code': campaign.code,
-                  'first-name': organizationLearner.firstName,
-                  'last-name': organizationLearner.lastName,
-                  birthdate: organizationLearner.birthdate,
-                  password: 'P@ssw0rd',
-                  username,
-                  'with-username': true,
-                },
-              },
-            },
-          });
-
-          // then
-          expect(response.statusCode).to.equal(422);
-          expect(response.result.errors[0].detail).to.equal(
-            'Cet identifiant n’est plus disponible, merci de recharger la page.'
-          );
-        });
-      });
-    });
-  });
-
   describe('POST /api/schooling-registration-dependent-users/generate-username-password', function () {
     let organizationId;
     let organizationLearnerId;
