+++ api/tests/integration/application/sco-organization-learners/index_test.js
@@ -16,6 +16,9 @@ describe('Integration | Application | Route | sco-organization-learners', functi
     sinon
       .stub(scoOrganizationLearnerController, 'generateUsername')
       .callsFake((request, h) => h.response('ok').code(200));
+    sinon
+      .stub(scoOrganizationLearnerController, 'createAndReconcileUserToOrganizationLearner')
+      .callsFake((request, h) => h.response().code(204));
 
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
@@ -830,6 +833,424 @@ describe('Integration | Application | Route | sco-organization-learners', functi
 
       // then
       expect(response.statusCode).to.equal(422);
+    });
+  });
+
+  describe('POST /api/schooling-registration-dependent-users', function () {
+    let method;
+    let url;
+    let payload;
+    let response;
+
+    context('When registration succeed with email', function () {
+      beforeEach(async function () {
+        // given
+        method = 'POST';
+        url = '/api/schooling-registration-dependent-users';
+        payload = {
+          data: {
+            attributes: {
+              'campaign-code': 'RESTRICTD',
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              email: 'robert.smith@example.net',
+              password: 'P@ssw0rd',
+              'with-username': false,
+              username: null,
+            },
+          },
+        };
+      });
+
+      it('should return 204', async function () {
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+    });
+
+    context('When registration succeed with username', function () {
+      beforeEach(async function () {
+        // given
+        method = 'POST';
+        url = '/api/schooling-registration-dependent-users';
+        payload = {
+          data: {
+            attributes: {
+              'campaign-code': 'RESTRICTD',
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              username: 'robert.smith1212',
+              password: 'P@ssw0rd',
+              'with-username': true,
+              email: null,
+            },
+          },
+        };
+      });
+
+      it('should return 204', async function () {
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+    });
+
+    context('Error cases', function () {
+      beforeEach(async function () {
+        // given
+        method = 'POST';
+        url = '/api/schooling-registration-dependent-users';
+        payload = {
+          data: {
+            attributes: {
+              'campaign-code': 'RESTRICTD',
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              username: 'robert.smith1212',
+              password: 'P@ssw0rd',
+              'with-username': true,
+            },
+          },
+        };
+      });
+
+      it('should return 400 when firstName is empty', async function () {
+        // given
+        payload.data.attributes['first-name'] = '';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when lastName is empty', async function () {
+        // given
+        payload.data.attributes['last-name'] = '';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when birthDate is not a valid date', async function () {
+        // given
+        payload.data.attributes.birthdate = '2012*-12-12';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when campaignCode is empty', async function () {
+        // given
+        payload.data.attributes['campaign-code'] = '';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when password is not valid', async function () {
+        // given
+        payload.data.attributes.password = 'not_valid';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when withUsername is not a boolean', async function () {
+        // given
+        payload.data.attributes['with-username'] = 'not_a_boolean';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      context('when username is not valid', function () {
+        it('should return 400 when username is an email', async function () {
+          // given
+          payload.data.attributes.username = 'robert.smith1212@example.net';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username has not dot between names', async function () {
+          // given
+          payload.data.attributes.username = 'robertsmith1212';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username does not end with 4 digits', async function () {
+          // given
+          payload.data.attributes.username = 'robert.smith';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username is capitalized', async function () {
+          // given
+          payload.data.attributes.username = 'Robert.Smith1212';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username is a phone number', async function () {
+          // given
+          payload.data.attributes.username = '0601010101';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+      });
+    });
+  });
+
+  describe('POST /api/sco-organization-learners/dependent', function () {
+    let method;
+    let url;
+    let payload;
+    let response;
+
+    context('When registration succeed with email', function () {
+      beforeEach(async function () {
+        // given
+        method = 'POST';
+        url = '/api/sco-organization-learners/dependent';
+        payload = {
+          data: {
+            attributes: {
+              'campaign-code': 'RESTRICTD',
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              email: 'robert.smith@example.net',
+              password: 'P@ssw0rd',
+              'with-username': false,
+              username: null,
+            },
+          },
+        };
+      });
+
+      it('should return 204', async function () {
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+    });
+
+    context('When registration succeed with username', function () {
+      beforeEach(async function () {
+        // given
+        method = 'POST';
+        url = '/api/sco-organization-learners/dependent';
+        payload = {
+          data: {
+            attributes: {
+              'campaign-code': 'RESTRICTD',
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              username: 'robert.smith1212',
+              password: 'P@ssw0rd',
+              'with-username': true,
+              email: null,
+            },
+          },
+        };
+      });
+
+      it('should return 204', async function () {
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+    });
+
+    context('Error cases', function () {
+      beforeEach(async function () {
+        // given
+        method = 'POST';
+        url = '/api/sco-organization-learners/dependent';
+        payload = {
+          data: {
+            attributes: {
+              'campaign-code': 'RESTRICTD',
+              'first-name': 'Robert',
+              'last-name': 'Smith',
+              birthdate: '2012-12-12',
+              username: 'robert.smith1212',
+              password: 'P@ssw0rd',
+              'with-username': true,
+            },
+          },
+        };
+      });
+
+      it('should return 400 when firstName is empty', async function () {
+        // given
+        payload.data.attributes['first-name'] = '';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when lastName is empty', async function () {
+        // given
+        payload.data.attributes['last-name'] = '';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when birthDate is not a valid date', async function () {
+        // given
+        payload.data.attributes.birthdate = '2012*-12-12';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when campaignCode is empty', async function () {
+        // given
+        payload.data.attributes['campaign-code'] = '';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when password is not valid', async function () {
+        // given
+        payload.data.attributes.password = 'not_valid';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      it('should return 400 when withUsername is not a boolean', async function () {
+        // given
+        payload.data.attributes['with-username'] = 'not_a_boolean';
+
+        // when
+        response = await httpTestServer.request(method, url, payload);
+
+        // then
+        expect(response.statusCode).to.equal(400);
+      });
+
+      context('when username is not valid', function () {
+        it('should return 400 when username is an email', async function () {
+          // given
+          payload.data.attributes.username = 'robert.smith1212@example.net';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username has not dot between names', async function () {
+          // given
+          payload.data.attributes.username = 'robertsmith1212';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username does not end with 4 digits', async function () {
+          // given
+          payload.data.attributes.username = 'robert.smith';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username is capitalized', async function () {
+          // given
+          payload.data.attributes.username = 'Robert.Smith1212';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+
+        it('should return 400 when username is a phone number', async function () {
+          // given
+          payload.data.attributes.username = '0601010101';
+
+          // when
+          response = await httpTestServer.request(method, url, payload);
+          // then
+          expect(response.statusCode).to.equal(400);
+        });
+      });
     });
   });
 });
