+++ api/tests/unit/application/sco-organization-learners/sco-organization-learner-controller_test.js
@@ -135,4 +135,46 @@ describe('Unit | Application | Controller | sco-organization-learner', function 
       expect(response.headers['Deprecation']).to.not.exist;
     });
   });
+
+  describe('#createAndReconcileUserToOrganizationLearner', function () {
+    const userId = 2;
+    const request = {
+      auth: { credentials: { userId } },
+      payload: { data: { attributes: {} } },
+    };
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'createAndReconcileUserToOrganizationLearner');
+      usecases.createAndReconcileUserToOrganizationLearner.resolves();
+    });
+
+    it('should return information about deprecation when old route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/schooling-registration-dependent-users',
+      };
+      const response = await scoOrganizationLearnerController.createAndReconcileUserToOrganizationLearner(
+        request,
+        hFake
+      );
+
+      // then
+      expect(response.headers['Deprecation']).to.equal('true');
+      expect(response.headers['Link']).to.equal('/api/sco-organization-learners/dependent; rel="successor-version"');
+    });
+
+    it('should not return information about deprecation when new route is used', async function () {
+      // when
+      hFake.request = {
+        path: '/api/sco-organization-learners/dependent',
+      };
+      const response = await scoOrganizationLearnerController.createAndReconcileUserToOrganizationLearner(
+        request,
+        hFake
+      );
+
+      // then
+      expect(response.headers['Deprecation']).to.not.exist;
+    });
+  });
 });
